MAINT: combine submodules, this isn't trivial to generalize
